// WaterBill/src/app/models/BillTransaction.ts
export class BillTransaction {
    BillId: any;
    GroupId: any;
    GroupName: any;
    BunglowNo: any;
    OwnerId: any;
    OwnerName: any;
    Year: any;
    Month: any;
    Maintenance: any;
    OtherCharge: any;
    Penalty: any;
    Total: any;
    LastUnit: any;
    CurrentUnit: any;
    PaymentType: any;
    BankName: any;
    BranchName: any;
    ChequeNo: any;
    ChequeDate: any;
    BillStatus: any;
    IsActive: any;
    StartDate: any;
    EndDate: any;
    ReceiptURL:any;
}